attempt to fix zero-sized entries on Windows for libarchive entries
On Windows, the size value provided via the stat struct of the
archive_entry_stat() function call seems to be zero in all cases.
To work around that issue the size is now retrieved directly via
archive_entry_size() instead. This should not make a difference
on Linux systems, but it can potentially fix the problem on
Windows systems.